pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "kgrag-store"
version = "1.0.0"
description = "A Python library for knowledge graph management and retrieval with advanced memory capabilities."
authors = [
    { name="Giuseppe Zileni", email="giuseppe.zileni@gmail.com" }
]
readme = {file = "README.md", content-type = "text/markdown"}
requires-python = ">=3.9"
license-files = ["LICENSE.md"]
keywords = ["agent", "knowledge", "memory", "retrieval", "vector-store", "kgrag"]
classifiers = ['Programming Language :: Python :: 3', 'Operating System :: OS Independent']
dependencies = [
    "aioboto3>=14.3.0",
    "aiobotocore>=2.22.0",
    "aiofiles>=24.1.0",
    "aiohappyeyeballs>=2.6.1",
    "aiohttp>=3.12.12",
    "aioitertools>=0.12.0",
    "aiosignal>=1.3.2",
    "annotated-types>=0.7.0",
    "anyio>=4.9.0",
    "attrs>=25.3.0",
    "backoff>=2.2.1",
    "beautifulsoup4>=4.13.4",
    "boto3>=1.37.3",
    "botocore>=1.37.3",
    "build>=1.2.2.post1",
    "certifi>=2025.4.26",
    "cffi>=1.17.1",
    "chardet>=5.2.0",
    "charset-normalizer>=3.4.2",
    "click>=8.2.1",
    "coloredlogs>=15.0.1",
    "contourpy>=1.3.2",
    "cryptography>=43.0.0",
    "cycler>=0.12.1",
    "dataclasses-json>=0.6.7",
    "distro>=1.9.0",
    "docutils>=0.21.2",
    "emoji>=2.14.1",
    "fastembed>=0.7.0",
    "filelock>=3.18.0",
    "filetype>=1.2.0",
    "flatbuffers>=25.2.10",
    "fonttools>=4.58.4",
    "fpdf>=1.7.2",
    "frozenlist>=1.7.0",
    "fsspec>=2024.12.0",
    "greenlet>=3.2.3",
    "grpcio>=1.73.0",
    "h11>=0.16.0",
    "h2>=4.2.0",
    "hf-xet>=1.1.3",
    "hpack>=4.1.0",
    "html5lib>=1.1",
    "httpcore>=1.0.9",
    "httpx>=0.28.1",
    "httpx-sse>=0.4.0",
    "huggingface-hub>=0.33.0",
    "humanfriendly>=10.0",
    "hyperframe>=6.1.0",
    "id>=1.5.0",
    "idna>=3.10",
    "iniconfig>=2.1.0",
    "jaraco.classes>=3.4.0",
    "jaraco.context>=6.0.1",
    "jaraco.functools>=4.1.0",
    "jiter>=0.10.0",
    "jmespath>=1.0.1",
    "joblib>=1.5.1",
    "json_repair>=0.39.1",
    "jsonpatch>=1.33",
    "jsonpointer>=3.0.0",
    "keyring>=25.6.0",
    "kiwisolver>=1.4.8",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "langchain-core>=0.3.65",
    "langchain-ollama>=0.3.3",
    "langchain-openai>=0.3.23",
    "langchain-qdrant>=0.2.0",
    "langchain-text-splitters>=0.3.8",
    "langdetect>=1.0.9",
    "langgraph>=0.4.8",
    "langgraph-checkpoint>=2.0.26",
    "langgraph-prebuilt>=0.2.2",
    "langgraph-sdk>=0.1.70",
    "langsmith>=0.3.45",
    "loguru>=0.7.3",
    "loki-logger-handler>=1.1.2",
    "lxml>=5.4.0",
    "markdown-it-py>=3.0.0",
    "marshmallow>=3.26.1",
    "matplotlib>=3.10.3",
    "mdurl>=0.1.2",
    "mmh3>=5.1.0",
    "more-itertools>=10.7.0",
    "mpmath>=1.3.0",
    "multidict>=6.4.4",
    "mypy_extensions>=1.1.0",
    "narwhals>=1.43.1",
    "neo4j>=5.28.1",
    "neo4j-graphrag>=1.7.0",
    "nest-asyncio>=1.6.0",
    "nh3>=0.2.21",
    "nltk>=3.9.1",
    "numpy>=1.26.4",
    "olefile>=0.47",
    "ollama>=0.5.1",
    "onnxruntime>=1.22.0",
    "openai>=1.86.0",
    "orjson>=3.10.18",
    "ormsgpack>=1.10.0",
    "packaging>=24.2",
    "pandas>=2.3.0",
    "pillow>=11.2.1",
    "plotly>=6.1.2",
    "pluggy>=1.6.0",
    "portalocker>=2.10.1",
    "propcache>=0.3.2",
    "protobuf>=6.31.1",
    "psutil>=7.0.0",
    "py_rust_stemmers>=0.1.5",
    "pyaws-s3>=1.0.21",
    "pycparser>=2.22",
    "pydantic>=2.11.6",
    "pydantic-settings>=2.9.1",
    "pydantic_core>=2.33.2",
    "Pygments>=2.19.1",
    "pyparsing>=3.2.3",
    "pypdf>=5.6.0",
    "pyproject_hooks>=1.2.0",
    "pytest>=8.4.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.0",
    "python-iso639>=2025.2.18",
    "python-magic>=0.4.27",
    "python-oxmsg>=0.0.2",
    "pytz>=2025.2",
    "PyYAML>=6.0.2",
    "qdrant-client>=1.14.2",
    "RapidFuzz>=3.13.0",
    "readme_renderer>=44.0",
    "redis>=6.2.0",
    "regex>=2024.11.6",
    "reportlab>=4.4.2",
    "requests>=2.32.4",
    "requests-toolbelt>=1.0.0",
    "rfc3986>=2.0.0",
    "rich>=14.0.0",
    "s3transfer>=0.11.3",
    "setuptools>=80.9.0",
    "six>=1.17.0",
    "sniffio>=1.3.1",
    "soupsieve>=2.7",
    "SQLAlchemy>=2.0.41",
    "sympy>=1.14.0",
    "tenacity>=9.1.2",
    "tiktoken>=0.9.0",
    "tokenizers>=0.21.1",
    "tqdm>=4.67.1",
    "twine>=6.1.0",
    "types-PyYAML>=6.0.12.20250516",
    "typing-inspect>=0.9.0",
    "typing-inspection>=0.4.1",
    "typing_extensions>=4.14.0",
    "tzdata>=2025.2",
    "unstructured>=0.17.2",
    "unstructured-client>=0.36.0",
    "urllib3>=2.4.0",
    "webencodings>=0.5.1",
    "wheel>=0.45.1",
    "wrapt>=1.17.2",
    "xxhash>=3.5.0",
    "yarl>=1.20.1",
    "zstandard>=0.23.0"
]

[project.urls]
Homepage = "https://gzileni.github.io/memory-agent"
Repository = "https://github.com/gzileni/memory-agent"

[tool.setuptools]
packages = ["kgrag_store"]  
